create ldap convenience function
