+++ tests/grippeweb/test_extract.py
@@ -2,7 +2,7 @@ from uuid import UUID
 
 import pytest
 
-from mex.common.ldap.models import LDAPActor, LDAPPerson
+from mex.common.ldap.models import LDAPFunctionalAccount, LDAPPerson
 from mex.common.models import AccessPlatformMapping, ResourceMapping
 from mex.common.types import MergedOrganizationIdentifier
 from mex.extractors.grippeweb.extract import (
@@ -45,12 +45,12 @@ def test_extract_ldap_actors_for_functional_accounts(
     ldap_actors = extract_ldap_actors_for_functional_accounts(
         grippeweb_resource_mappings
     )
-    expected = LDAPActor(
+    expected = LDAPFunctionalAccount(
+        objectGUID=UUID("00000000-0000-4000-8000-000000000004"),
         sAMAccountName="ContactC",
-        objectGUID=UUID(int=4, version=4),
         mail=["email@email.de", "contactc@rki.de"],
+        ou=["Funktion"],
     )
-
     assert ldap_actors[0] == expected
