+++ tests/igs/test_extract.py
@@ -2,7 +2,7 @@ from uuid import UUID
 
 import pytest
 
-from mex.common.ldap.models import LDAPActor
+from mex.common.ldap.models import LDAPFunctionalAccount
 from mex.common.models import AccessPlatformMapping, ResourceMapping
 from mex.extractors.igs.extract import (
     extract_igs_schemas,
@@ -34,15 +34,17 @@ def test_extract_ldap_actors_by_mail(
         igs_resource_mapping, igs_access_platform_mapping
     )
     expected = {
-        "contactc@rki.de": LDAPActor(
-            sAMAccountName="ContactC",
+        "contactc@rki.de": LDAPFunctionalAccount(
             objectGUID=UUID("00000000-0000-4000-8000-000000000004"),
+            sAMAccountName="ContactC",
             mail=["email@email.de", "contactc@rki.de"],
+            ou=["Funktion"],
         ),
-        "email@email.de": LDAPActor(
-            sAMAccountName="ContactC",
+        "email@email.de": LDAPFunctionalAccount(
             objectGUID=UUID("00000000-0000-4000-8000-000000000004"),
+            sAMAccountName="ContactC",
             mail=["email@email.de", "contactc@rki.de"],
+            ou=["Funktion"],
         ),
     }
     assert ldap_actors == expected
