+++ tests/synopse/test_extract.py
@@ -102,9 +102,10 @@ def test_extract_synopse_contact(
 ) -> None:
     actor = extract_synopse_contact(synopse_access_platform)
     expected = {
-        "sAMAccountName": "ContactC",
         "objectGUID": UUID("00000000-0000-4000-8000-000000000004"),
+        "sAMAccountName": "ContactC",
         "mail": ["email@email.de", "contactc@rki.de"],
+        "ou": ["Funktion"],
     }
     assert actor[0].model_dump(exclude_none=True) == expected
